requirements.txt
requests
pandas
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-auth-httplib2
pyarrow
python-dotenv
apache-airflow